+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - add more entries to the credits modal
 - update remove all accessories component
 - improvements to matter components
+- feat: air quality sensor matter device
 
 ### Other Changes